test(sync): isolate screenshot ordering coverage
